fix tuxedo_keyboard installation by using their RPM package
See their new repo here :
https://gitlab.com/tuxedocomputers/development/packages/tuxedo-drivers .

The old tuxedo-keyboard repo (
https://github.com/tuxedocomputers/tuxedo-keyboard  ) is archived, and
has been replaced by the new one above.

Also, they now distribute a .rpm of their tuxedo-drivers package (which
contains the keyboard backlight driver).

So we now install the keyboad backlight driver via the rpm as it is
easier instead of relying on sources clone and compiling. Also, the
package version upgrade is now done with dnf when upgrading the
installed packages (as for any other packages ;) ).